Simplify `hasFormMutator` method
The previous implementation that relied on Reflection is causing fatal PHP errors, when a Model which uses the FormAccessible trait is being dumped (with Symfony's dumper).

The reason is because the dumper is using reflection to dump all the model's properties, and when it hits the `reflection` property, it tries to cast the `ReflectionClass` using `ReflectionClass`, which causes an `Internal error: Failed to retrieve the reflection object`.

The new implementation uses the same stragegy as Eloquent's `hasGetMutator`, which calls `method_exists`.

This is more efficient as we don't need a whole reflection object, neither we need to iterate through all the reflected public properties of the model, just to find the first one that matches the mutator method name.