Pre-process date values
Given that a form is opened with Form::model
When using Form::date to create a date field and leaving the value null to fetch the value from the model, when the value is a Carbon then the value is set as Y-m-d H:i:s and the appropriate date is not selected. If you pass a carbon directly to the value (instead of null) then it is correctly formatted to Y-m-d.

This pull request pre-processes the value when it's null using getValueAttribute. Additionally, it replaces DateTime with DateTimeInterface to allow the use of CarbonImmutable